[MM-36009]Fixes:CRT: Shortcut (Command-/) for the keyboard shorcuts Modal doesn't work in global threads
